Revert "GThread: Don't g_error() if setting the thread scheduler settings fails"
This reverts commit 965061797d74847d2e2d1237bb7a63328608e28c.

We are having trouble tracking down the cause of #2769. When the bug
occurs, we fail to set scheduler settings for the new thread pool
thread. This can have serious consequences and should not be ignored. In
retrospect, making this a critical instead of a fatal error has made it
more difficult to notice, debug, and fix. This operation needs to always
work, so let's crash when it fails.

This does not fix #2769, but will hopefully help.