gmessages: Add support for G_MESSAGES_USE_STDERR env variable
Make possible to run glib processes where all the GMessage's output is
written to stderr.

By default we do it only for log levels greater or equal than warning,
but in some cases we may want to send all the glib logging to stderr.

This was possible to do programmatically, but there might be still cases
in which this has to be setup in the environment.

So, add support for a new environment variable, without changing the
ability for a program to set this to another value (at early start).