po: Add missing header file to POTFILES.in
Change originally by Piotr in !3257, but the CI has problems with
his name in the author field. Since this change is blocking a release,
I have decided to merge it under my name instead, so that it lands sooner.

In parallel, we are trying to figure out what is broken with the CI so
this does not happen again.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>