ci: Ensure we don't accidentally use too-new macOS API without guards
On macOS, when building with a new SDK and targeting an older version,
say, macOS High Sierra (95% market share), you need to either not use
new APIs, or you need to use the `__builtin_available` compiler
built-in to ensure that your code is loaded conditionally.

For example:

```c
if (__builtin_available (macOS 10.13, *)) {
  <code that uses API added in 10.13>
} else {
  <old code goes here>
}
```

You can also combine it with a compiler check for whether the SDK
supports it like so:

```
 #if defined(MAC_OS_VERSION_11_0) && MAC_OS_X_VERSION_MAX_ALLOWED >= MAC_OS_VERSION_11_0
  if (__builtin_available (macOS 11.00, *)) {
    <code that uses API added in 11.00>
  }
 #endif
```

Without this, code will compile file but crash at runtime while trying
to lazy-load the symbol.